requirements.txt
pip==20.0.2
Django==3.0.3
djangorestframework==3.11.0
psycopg2==2.8.4
django-extensions==2.2.7
django-redis==4.11.0